refactor: simplify SSTable builder and tests
